Work in progress for a tkinter testGUI
This GUI will test commands,
Running however back to the problem in the old days of executing an
application logic in Dragonfly, so will need to use a subprocess again
as like is done for engine. Big change that then the Show test command
window command will come in Dragonall if that is realized. Also
TestCommandGui should be placed then in a lib folder, as should Engine
then. So will pick this work up again.